feat: Add static server for serving static assets
This commit adds a static server to the app that can be used to serve static assets, such as images, CSS, and JavaScript files. The server is started on port 8080 and can be accessed by visiting http://localhost:8080/index.html.

This feature will allow us to serve static assets from the app itself, which will improve performance and reduce the need for external dependencies.